requirements.txt
appnope==0.1.0
attrs==19.3.0
beautifulsoup4==4.9.0
bleach==3.1.4
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.4.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.1
idna==2.8
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
lxml==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
notebook==5.2.2
numpy==1.18.2
oauthlib==3.1.0
pandas==1.0.3
pandocfilters==1.4.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==1.0.18
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
simplegeneric==0.8.1
six==1.12.0
soupsieve==2.0
terminado==0.8.3
testpath==0.4.4
tornado==4.5.3
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.24.3
wcwidth==0.1.9
webencodings==0.5.1
XlsxWriter==1.2.8
